+++ widgets/linechart/linechart_test.go
@@ -16,6 +16,7 @@ package linechart
 
 import (
 	"image"
+	"math"
 	"testing"
 
 	"github.com/kylelemons/godebug/pretty"
@@ -36,9 +37,42 @@ func TestLineChartDraws(t *testing.T) {
 		opts         []Option
 		writes       func(*LineChart) error
 		want         func(size image.Point) *faketerm.Terminal
-		wantWriteErr bool
 		wantErr      bool
+		wantWriteErr bool
+		wantDrawErr  bool
 	}{
+		{
+			desc:   "fails with custom scale where min is NaN",
+			canvas: image.Rect(0, 0, 3, 4),
+			opts: []Option{
+				YAxisCustomScale(math.NaN(), 1),
+			},
+			wantErr: true,
+		},
+		{
+			desc:   "fails with custom scale where max is NaN",
+			canvas: image.Rect(0, 0, 3, 4),
+			opts: []Option{
+				YAxisCustomScale(0, math.NaN()),
+			},
+			wantErr: true,
+		},
+		{
+			desc:   "fails with custom scale where min > max",
+			canvas: image.Rect(0, 0, 3, 4),
+			opts: []Option{
+				YAxisCustomScale(1, 0),
+			},
+			wantErr: true,
+		},
+		{
+			desc:   "fails with custom scale where min == max",
+			canvas: image.Rect(0, 0, 3, 4),
+			opts: []Option{
+				YAxisCustomScale(1, 1),
+			},
+			wantErr: true,
+		},
 		{
 			desc:   "series fails without name for the series",
 			canvas: image.Rect(0, 0, 3, 4),
@@ -177,6 +211,187 @@ func TestLineChartDraws(t *testing.T) {
 				// Value labels.
 				testdraw.MustText(c, "0", image.Point{4, 7})
 				testdraw.MustText(c, "51.68", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{6, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(6, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{26, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc: "custom Y scale, zero based positive, values fit",
+			opts: []Option{
+				YAxisCustomScale(0, 200),
+			},
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{6, 0}, End: image.Point{6, 8}},
+					{Start: image.Point{6, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "0", image.Point{5, 7})
+				testdraw.MustText(c, "103.36", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{7, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(7, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{25, 16})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc: "custom Y scale, zero based negative, values fit",
+			opts: []Option{
+				YAxisCustomScale(-200, 0),
+			},
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, -200})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{6, 0}, End: image.Point{6, 8}},
+					{Start: image.Point{6, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "-200", image.Point{2, 7})
+				testdraw.MustText(c, "-96.64", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{7, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(7, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 0}, image.Point{25, 31})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc: "custom Y scale, negative and positive, values fit",
+			opts: []Option{
+				YAxisCustomScale(-200, 200),
+			},
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{0, 100})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{4, 0}, End: image.Point{4, 8}},
+					{Start: image.Point{4, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "-200", image.Point{0, 7})
+				testdraw.MustText(c, "6.57", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{5, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(5, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 16}, image.Point{29, 8})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc: "custom Y scale, negative only, values fit",
+			opts: []Option{
+				YAxisCustomScale(-200, -100),
+				YAxisAdaptive(),
+			},
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{-200, -100})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{7, 0}, End: image.Point{7, 8}},
+					{Start: image.Point{7, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "-200", image.Point{3, 7})
+				testdraw.MustText(c, "-148.32", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{8, 9})
+				testdraw.MustText(c, "1", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(8, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{23, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc: "custom Y scale, negative and positive, values don't fit so adjusted",
+			opts: []Option{
+				YAxisCustomScale(-200, 200),
+			},
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				return lc.Series("first", []float64{-400, 400})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{5, 0}, End: image.Point{5, 8}},
+					{Start: image.Point{5, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "-400", image.Point{1, 7})
+				testdraw.MustText(c, "12.96", image.Point{0, 3})
 				testdraw.MustText(c, "0", image.Point{6, 9})
 				testdraw.MustText(c, "1", image.Point{19, 9})
 
@@ -548,6 +763,84 @@ func TestLineChartDraws(t *testing.T) {
 				return ft
 			},
 		},
+		{
+			desc:   "draw multiple series, the second has smaller scale than the first",
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				if err := lc.Series("first", []float64{0, 50, 100}); err != nil {
+					return err
+				}
+				return lc.Series("second", []float64{10, 20})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{5, 0}, End: image.Point{5, 8}},
+					{Start: image.Point{5, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "0", image.Point{4, 7})
+				testdraw.MustText(c, "51.68", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{6, 9})
+				testdraw.MustText(c, "1", image.Point{12, 9})
+				testdraw.MustText(c, "2", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(6, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{13, 16})
+				testdraw.MustBrailleLine(bc, image.Point{13, 16}, image.Point{27, 0})
+				testdraw.MustBrailleLine(bc, image.Point{0, 28}, image.Point{13, 25})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
+		{
+			desc:   "draw multiple series, the second has larger scale than the first",
+			canvas: image.Rect(0, 0, 20, 10),
+			writes: func(lc *LineChart) error {
+				if err := lc.Series("first", []float64{0, 50, 100}); err != nil {
+					return err
+				}
+				return lc.Series("second", []float64{-10, 200})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				c := testcanvas.MustNew(ft.Area())
+
+				// Y and X axis.
+				lines := []draw.HVLine{
+					{Start: image.Point{5, 0}, End: image.Point{5, 8}},
+					{Start: image.Point{5, 8}, End: image.Point{19, 8}},
+				}
+				testdraw.MustHVLines(c, lines)
+
+				// Value labels.
+				testdraw.MustText(c, "-10", image.Point{2, 7})
+				testdraw.MustText(c, "98.48", image.Point{0, 3})
+				testdraw.MustText(c, "0", image.Point{6, 9})
+				testdraw.MustText(c, "1", image.Point{12, 9})
+				testdraw.MustText(c, "2", image.Point{19, 9})
+
+				// Braille line.
+				graphAr := image.Rect(6, 0, 20, 8)
+				bc := testbraille.MustNew(graphAr)
+				testdraw.MustBrailleLine(bc, image.Point{0, 30}, image.Point{13, 22})
+				testdraw.MustBrailleLine(bc, image.Point{13, 22}, image.Point{27, 15})
+				testdraw.MustBrailleLine(bc, image.Point{0, 31}, image.Point{13, 0})
+				testbraille.MustCopyTo(bc, c)
+
+				testcanvas.MustApply(c, ft)
+				return ft
+			},
+		},
 	}
 
 	for _, tc := range tests {
@@ -557,7 +850,14 @@ func TestLineChartDraws(t *testing.T) {
 				t.Fatalf("canvas.New => unexpected error: %v", err)
 			}
 
-			widget := New(tc.opts...)
+			widget, err := New(tc.opts...)
+			if (err != nil) != tc.wantErr {
+				t.Errorf("New => unexpected error: %v, wantErr: %v", err, tc.wantErr)
+			}
+			if err != nil {
+				return
+			}
+
 			if tc.writes != nil {
 				err := tc.writes(widget)
 				if (err != nil) != tc.wantWriteErr {
@@ -570,8 +870,8 @@ func TestLineChartDraws(t *testing.T) {
 
 			{
 				err := widget.Draw(c)
-				if (err != nil) != tc.wantErr {
-					t.Fatalf("Draw => unexpected error: %v, wantErr: %v", err, tc.wantErr)
+				if (err != nil) != tc.wantDrawErr {
+					t.Fatalf("Draw => unexpected error: %v, wantDrawErr: %v", err, tc.wantDrawErr)
 				}
 				if err != nil {
 					return
@@ -658,7 +958,10 @@ func TestOptions(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.desc, func(t *testing.T) {
-			lc := New(tc.opts...)
+			lc, err := New(tc.opts...)
+			if err != nil {
+				t.Fatalf("New => unexpected error: %v", err)
+			}
 
 			if tc.addSeries != nil {
 				if err := tc.addSeries(lc); err != nil {